requirements.txt
# Core ML and Data Science
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Computer Vision
opencv-python==4.8.0.74
Pillow==10.0.0
albumentations==1.3.1
imgaug==0.4.0

# DNA/Bioinformatics
biopython==1.81
pysam==0.21.0
pyfaidx==0.7.1
blast==0.0.1

# Web Framework
fastapi==0.101.1
uvicorn==0.23.2
pydantic==2.0.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.19
sqlite3
psycopg2-binary==2.9.7

# Image Processing
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1

# Deployment
docker==6.1.0
gunicorn==21.2.0
